requirements.txt
flask
cloudinary
python-dotenv
werkzeug
gunicorn
